Refactor: Move types into appropriate folders
