+++ client/laundroclean/src/services/laundrocleanservices/laundrocleanservices.service.ts
@@ -1,18 +1,14 @@
-import type { StaticImport } from "next/dist/shared/lib/get-img-props";
 
-export interface ServiceDisplayProp {
-  id?: string;
-  name: string;
-  description?: string;
-  pricingType: string;
-  amount: number;
-  currency: string;
-  maxDailyBookings?: string;
-}
 
-export interface ServicesProps {
-  icon?: string | StaticImport;
-  name: string;
-  description: string;
-  orderedlist?: string[];
-}
+
+
+// Hook
+/*export const useServices = () => {
+  return useQuery({
+    queryKey: ['services', 'promoCodes'],
+    queryFn: async () => {
+      const serviceResponse = await apiRequest<ServicesPromoResponse>('services');
+      // const promo
+    }
+  })
+} */
